废弃了CTAppContext，主观觉得这是网络层的封装，他提供的内容太多，降低其权限内容，增加CTNetworkingConfigurationManager做其补充，废弃CTUDIDGenerator，同样主观觉得这部分不属于能用逻辑
