+++ CTNetworking/CTNetworking/CTNetworking.h
@@ -11,8 +11,7 @@
 
 #import "CTApiProxy.h"
 #import "CTServiceFactory.h"
-#import "CTAppContext.h"
 #import "CTAPIBaseManager.h"
-#import "CTNetworkingConfiguration.h"
+#import "CTNetworkingConfigurationManager.h"
 
 #endif
