+++ CTNetworking/CTNetworking/Categories/UIDevice+CTNetworkingMethods.h
@@ -16,20 +16,9 @@
  * @description apple identifier support iOS6 and iOS5 below
  */
 
-- (NSString *) CT_uuid;
-- (NSString *) CT_udid;
 - (NSString *) CT_macaddress;
 - (NSString *) CT_macaddressMD5;
 - (NSString *) CT_machineType;
 - (NSString *) CT_ostype;//显示“ios6，ios5”，只显示大版本号
-- (NSString *) CT_createUUID;
 
-//兼容旧版本
-- (NSString *) uuid;
-- (NSString *) udid;
-- (NSString *) macaddress;
-- (NSString *) macaddressMD5;
-- (NSString *) machineType;
-- (NSString *) ostype;//显示“ios6，ios5”，只显示大版本号
-- (NSString *) createUUID;
 @end
