+++ CTNetworking/CTNetworking/Components/URLResponse/CTURLResponse.h
@@ -7,7 +7,13 @@
 //
 
 #import <Foundation/Foundation.h>
-#import "CTNetworkingConfiguration.h"
+
+typedef NS_ENUM(NSUInteger, CTURLResponseStatus)
+{
+    CTURLResponseStatusSuccess, //作为底层，请求是否成功只考虑是否成功收到服务器反馈。至于签名是否正确，返回的数据是否完整，由上层的CTAPIBaseManager来决定。
+    CTURLResponseStatusErrorTimeout,
+    CTURLResponseStatusErrorNoNetwork // 默认除了超时以外的错误都是无网络错误。
+};
 
 @interface CTURLResponse : NSObject
