+++ CTNetworking/CTNetworkingCustoms/Configurations/CTNetworkingConfigurationManager.h
@@ -1,0 +1,23 @@
+//
+//  CTNetworkingConfigurationManager.h
+//  CTNetworking
+//
+//  Created by Corotata on 2017/4/10.
+//  Copyright © 2017年 Long Fan. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface CTNetworkingConfigurationManager : NSObject
+
++ (instancetype)sharedInstance;
+
+@property (nonatomic, assign, readonly) BOOL isReachable;
+
+@property (nonatomic, assign) BOOL shouldCache;
+@property (nonatomic, assign) BOOL serviceIsOnline;
+@property (nonatomic, assign) NSTimeInterval apiNetworkingTimeoutSeconds;
+@property (nonatomic, assign) NSTimeInterval cacheOutdateTimeSeconds;
+@property (nonatomic, assign) NSInteger cacheCountLimit;
+
+@end
