+++ CTNetworking/TestCase/CTHTTPConst.h
@@ -1,0 +1,11 @@
+//
+//  CTHTTPConst.h
+//  CTNetworking
+//
+//  Created by Corotata on 2017/4/10.
+//  Copyright © 2017年 Long Fan. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+extern  NSString * const kCTServiceGDMapV3;
