+++ CTNetworking/TestCase/CTHTTPConst.m
@@ -1,0 +1,11 @@
+//
+//  CTHTTPConst.m
+//  CTNetworking
+//
+//  Created by Corotata on 2017/4/10.
+//  Copyright © 2017年 Long Fan. All rights reserved.
+//
+
+#import "CTHTTPConst.h"
+
+NSString * const kCTServiceGDMapV3 = @"kCTServiceGDMapV3";
